feat: Quiz 사진 등록, 목록 조회 추가

+++ backend/src/main/java/com/edufocus/edufocus/quiz/service/QuizSetServiceImpl.java
@@ -1,7 +1,6 @@
 package com.edufocus.edufocus.quiz.service;
 
 import com.edufocus.edufocus.quiz.entity.*;
-import com.edufocus.edufocus.quiz.repository.QuizRepository;
 import com.edufocus.edufocus.quiz.repository.QuizSetRepository;
 import com.edufocus.edufocus.user.model.entity.User;
 import com.edufocus.edufocus.user.model.repository.UserRepository;
@@ -11,6 +10,7 @@ import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 
 @Service
 @Transactional
@@ -38,36 +38,65 @@ public class QuizSetServiceImpl implements QuizSetService {
     }
 
     @Override
-    public void deleteQuizSet(long quizSetId) {
+    public boolean deleteQuizSet(Long userId, Long quizSetId) {
+        Optional<QuizSet> quizSet = quizSetRepository.findById(quizSetId);
+
+        if (quizSet.isEmpty() || userId != quizSet.get().getUser().getId()) {
+            return false;
+        }
+
         quizSetRepository.deleteById(quizSetId);
+        return true;
     }
 
     @Override
-    public QuizSet findQuizSet(long quizSetId) {
+    public QuizSet findQuizSet(Long quizSetId) {
         return quizSetRepository.findById(quizSetId).get();
     }
 
     @Override
-    public QuizSetResponse findQuizSetResponse(long quizSetId) {
-        QuizSet quizSet = findQuizSet(quizSetId);
+    public QuizSetResponse findQuizSetResponse(Long quizSetId) {
+        Optional<QuizSet> quizSet = quizSetRepository.findById(quizSetId);
+
+        if (quizSet.isEmpty()) {
+            return null;
+        }
+
 
         List<QuizResponse> quizResponses = new ArrayList<>();
-        for (Quiz quiz : quizSet.getQuizzes()) {
+        for (Quiz quiz : quizSet.get().getQuizzes()) {
             QuizResponse quizResponse = new QuizResponse().builder()
                     .question(quiz.getQuestion())
+                    .image(quiz.getImage())
                     .choices(quiz.getChoices())
                     .build();
             quizResponses.add(quizResponse);
         }
 
         QuizSetResponse quizSetResponse = new QuizSetResponse().builder()
-                .title(quizSet.getTitle())
-                .image(quizSet.getImage())
+                .title(quizSet.get().getTitle())
                 .quizzes(quizResponses)
                 .build();
 
         return quizSetResponse;
     }
 
+    @Override
+    public List<MyQuizSetResponse> findMyQuizSetResponses(Long userId) {
+        List<QuizSet> quizSetList = quizSetRepository.findQuizSetsByUserId(userId);
+
+        List<MyQuizSetResponse> myQuizSetResponses = new ArrayList<>();
+        for (QuizSet quizSet : quizSetList) {
+            MyQuizSetResponse myQuizSetResponse = new MyQuizSetResponse().builder()
+                    .quizSetId(quizSet.getId())
+                    .title(quizSet.getTitle())
+                    .build();
+
+            myQuizSetResponses.add(myQuizSetResponse);
+        }
+
+        return myQuizSetResponses;
+
+    }
 
 }
